Actually run at 60fps
Using `setInterval()` with a 16ms interval won't actually cause the progress bar to animate at 60fps. If you collect a profile using Chrome DevTools you'll notice that it misses a lot of frames as the interval gets out of sync with the display refresh. Using `requestAnimationFrame()` instead causes the browser to update the current progress immediately before rendering the next frame, making it update at whatever rate the system is running at.